+++ client/src/rpc/types/storage_root.rs
@@ -2,7 +2,7 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-use crate::rpc::types::H256;
+use cfx_types::H256;
 use primitives::StorageRoot as PrimitiveStorageRoot;
 
 #[derive(Debug, Serialize, Deserialize)]